+++ angular-workspace/projects/example-client-app/src/app/app.module.ts
@@ -1,7 +1,9 @@
-import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
+import { FormBuilder, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
-import { NimbleTextFieldModule, NimbleNumberFieldModule } from '@ni/nimble-angular';
+import { NimbleTextFieldModule, NimbleNumberFieldModule, NimbleSelectModule, NimbleListboxOptionModule, NimbleButtonModule } from '@ni/nimble-angular';
+import { NimbleThemeProviderModule } from 'projects/ni/nimble-angular/src/directives/theme-provider';
 
 import { AppComponent } from './app.component';
 import { LoginComponent } from './login/login.component';
@@ -13,16 +15,21 @@ import { LoginComponent } from './login/login.component';
     ],
     imports: [
         BrowserModule,
+        FormsModule,
+        ReactiveFormsModule,
+        NimbleThemeProviderModule,
         NimbleTextFieldModule,
         NimbleNumberFieldModule,
+        NimbleSelectModule,
+        NimbleListboxOptionModule,
+        NimbleButtonModule,
         RouterModule.forRoot([
             { path: '', redirectTo: '/login', pathMatch: 'full' },
             { path: 'login', component: LoginComponent }
         ],
         { useHash: true })
     ],
-    providers: [],
-    bootstrap: [AppComponent],
-    schemas: [CUSTOM_ELEMENTS_SCHEMA]
+    providers: [FormBuilder],
+    bootstrap: [AppComponent]
 })
 export class AppModule { }