+++ angular-workspace/projects/example-client-app/src/app/login/login.component.ts
@@ -1,4 +1,6 @@
 import { Component } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { ButtonAppearance } from '@ni/nimble-components/dist/esm/button/types';
 import { NimbleTheme } from '@ni/nimble-components/dist/esm/theme-provider/themes';
 
 @Component({
@@ -7,12 +9,15 @@ import { NimbleTheme } from '@ni/nimble-components/dist/esm/theme-provider/theme
     styleUrls: ['./login.component.scss']
 })
 export class LoginComponent {
+    public loginForm: FormGroup;
     public theme: NimbleTheme = NimbleTheme.Light;
     public themes = NimbleTheme;
-    private toggle = true;
+    public buttonAppearance = ButtonAppearance;
 
-    public switchThemes(): void {
-        this.toggle = !this.toggle;
-        this.theme = this.toggle ? NimbleTheme.Light : NimbleTheme.Dark;
+    public constructor(private readonly formBuilder: FormBuilder) {
+        this.loginForm = this.formBuilder.group({
+            username: [''],
+            password: ['', Validators.required]
+        });
     }
 }