+++ angular-workspace/projects/ni/nimble-angular/src/components/text-field/index.ts
@@ -1,2 +1,0 @@
-export * from './nimble-text-field.module';
-export * from './nimble-text-field.component';