+++ packages/nimble-components/src/number-field/styles.ts
@@ -4,11 +4,14 @@ import {
     borderColorHover,
     borderWidth,
     contentFontColorDisabled,
+    contentFontSize,
+    controlHeight,
     fillColorSelectedRgb,
     fontFamily,
     labelFontColor,
     labelFontFamily,
     labelFontSize,
+    labelHeight,
     labelTextTransform,
     smallDelay
 } from '../theme-provider/design-tokens';
@@ -17,9 +20,11 @@ export const styles = css`
     :host {
         display: inline-block;
         font-family: ${fontFamily};
+        font-size: ${contentFontSize};
         outline: none;
         user-select: none;
         color: ${labelFontColor};
+        height: calc(${labelHeight} + ${controlHeight});
     }
 
     :host([disabled]) {
@@ -35,7 +40,8 @@ export const styles = css`
         border-radius: 0px;
         font-family: ${fontFamily};
         border-bottom: ${borderWidth} solid ${borderColor};
-        transition: border-bottom ${smallDelay};
+        padding-bottom: 1px;
+        transition: border-bottom ${smallDelay}, padding-bottom ${smallDelay};
     }
 
     @media (prefers-reduced-motion) {
@@ -46,6 +52,7 @@ export const styles = css`
 
     .root:hover {
         border-bottom: 2px solid ${borderColorHover};
+        padding-bottom: 0px;
     }
 
     :host([disabled]) .root,
@@ -57,7 +64,6 @@ export const styles = css`
         -webkit-appearance: none;
         font: inherit;
         background: transparent;
-        border: 0;
         color: inherit;
         height: 28px;
         width: 100%;
@@ -80,8 +86,7 @@ export const styles = css`
 
     .control::placeholder {
         color: ${labelFontColor};
-        font-style: italic;
-        opacity: 0.5;
+        opacity: 0.6;
     }
 
     .control:focus-within::placeholder {
@@ -93,9 +98,10 @@ export const styles = css`
     }
 
     .label {
+        display: flex;
         font-family: ${labelFontFamily};
         font-size: ${labelFontSize};
-        line-height: 16px;
+        line-height: ${labelHeight};
         text-transform: ${labelTextTransform};
     }