+++ packages/nimble-components/src/tab-panel/index.ts
@@ -1,0 +1,14 @@
+import {
+    DesignSystem,
+    TabPanel,
+    tabPanelTemplate as template
+} from '@microsoft/fast-foundation';
+import { styles } from './styles';
+
+export const nimbleTabPanel = TabPanel.compose({
+    baseName: 'tab-panel',
+    template,
+    styles
+});
+
+DesignSystem.getOrCreate().withPrefix('nimble').register(nimbleTabPanel());