+++ packages/nimble-components/src/theme-provider/design-tokens.ts
@@ -9,12 +9,14 @@ import {
     White,
     Enterprise,
     Selection100,
-    BodyFamily,
-    OverlineCapitalizedFamily,
     SlLegacyBlue,
     Fail100LightUi,
     SmallDelay,
-    Fail100DarkUi
+    Fail100DarkUi,
+    BodyFamily,
+    BodySize,
+    OverlineCapitalizedFamily,
+    OverlineCapitalizedSize
 } from '@ni/nimble-tokens/dist/styledictionary/js/tokens';
 import { NimbleTheme } from './themes';
 
@@ -109,12 +111,22 @@ function getLabelTextSize(element: HTMLElement): string {
         case NimbleTheme.LegacyBlue:
             return '13px';
         default:
-            return '11px';
+            return OverlineCapitalizedSize;
+    }
+}
+
+function getPasswordRevealFilter(element: HTMLElement): string {
+    switch (theme.getValueFor(element)) {
+        case NimbleTheme.Color:
+        case NimbleTheme.Dark:
+            return 'invert(100%)';
+        default:
+            return 'invert(0%)';
     }
 }
 
 const fillColorSelectedTheme = (element: HTMLElement): string => getColorForTheme(element, Selection100, Selection100, White, SlLegacyBlue);
-const fillColorHoverTheme = (element: HTMLElement): string => getColorForTheme(element, Black91, Black15, White, Black91);
+const fillColorHoverTheme = (element: HTMLElement): string => getColorForTheme(element, Black91, Black15, White, SlLegacyBlue);
 
 // Color Tokens
 export const actionColorRgb = create<string>('action-color-rgb').withDefault(
@@ -171,9 +183,21 @@ export const borderColorHover = create<string>(
     hexToRgba(SlLegacyBlue, 0.9)
 ));
 
+export const popupBoxShadowColor = create<string>(
+    'popup-box-shadow-color'
+).withDefault((element: HTMLElement) => hexToRgba(
+    getColorForTheme(element, Black75, Black85, Black85, Black75),
+    0.3
+));
+
+export const popupBorderColor = create<string>(
+    'popup-border-color'
+).withDefault((element: HTMLElement) => hexToRgba(getColorForTheme(element, Black91, Black15, White, Black91), 0.3));
+
 // Component Sizing Tokens
 export const controlHeight = create<string>('control-height').withDefault('32px');
 export const standardPadding = create<string>('standard-padding').withDefault('16px');
+export const labelHeight = create<string>('label-height').withDefault('16px');
 export const borderWidth = create<string>('border-width').withDefault('1px');
 
 // Font Family Tokens
@@ -186,7 +210,7 @@ export const labelFontFamily = create<string>('label-font-family').withDefault(
 export const labelFontSize = create<string>('label-font-size').withDefault(
     (element: HTMLElement) => getLabelTextSize(element)
 );
-export const contentFontSize = create<string>('content-font-size').withDefault('14px');
+export const contentFontSize = create<string>('content-font-size').withDefault(BodySize);
 
 // Font Color Tokens
 export const labelFontColor = create<string>('label-font-color').withDefault(
@@ -215,3 +239,8 @@ export const contentFontColorDisabled = create<string>(
 
 // Animation Tokens
 export const smallDelay = create<string>('small-delay').withDefault(SmallDelay);
+
+// Filter Tokens
+export const passwordRevealFilter = create<string>(
+    'password-reveal-filter'
+).withDefault((element: HTMLElement) => getPasswordRevealFilter(element));