+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,139 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Thu, 02 Sep 2021 13:46:41 GMT and should not be manually modified.
+This log was last generated on Thu, 16 Sep 2021 15:30:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.0-beta.53
+
+Thu, 16 Sep 2021 15:30:58 GMT
+
+### Changes
+
+- Add simple tab component (9480438+LukasKall@users.noreply.github.com)
+
+## 1.0.0-beta.52
+
+Wed, 15 Sep 2021 20:07:53 GMT
+
+### Changes
+
+- fix bottom borders of text-field and number-field (22551874+scotia673@users.noreply.github.com)
+
+## 1.0.0-beta.51
+
+Tue, 14 Sep 2021 21:19:06 GMT
+
+### Changes
+
+- Use VSCode Tasks to start watch scripts (1458528+fredvisser@users.noreply.github.com)
+- Bump @ni/nimble-components to v1.0.0-beta.51 (1458528+fredvisser@users.noreply.github.com)
+
+## 1.0.0-beta.50
+
+Tue, 14 Sep 2021 16:53:39 GMT
+
+### Changes
+
+- uptake DSP font sizes in design-tokens.ts (22551874+scotia673@users.noreply.github.com)
+- Bump @ni/nimble-components to v1.0.0-beta.50 (1458528+fredvisser@users.noreply.github.com)
+
+## 1.0.0-beta.49
+
+Mon, 13 Sep 2021 15:09:27 GMT
+
+### Changes
+
+- Add karma reporter to list all tests run during CI (jattasNI@users.noreply.github.com)
+
+## 1.0.0-beta.48
+
+Fri, 10 Sep 2021 19:03:37 GMT
+
+### Changes
+
+- Fix broken link in CONTRIBUTING.md (jattasNI@users.noreply.github.com)
+
+## 1.0.0-beta.47
+
+Fri, 10 Sep 2021 18:13:15 GMT
+
+### Changes
+
+- Created nimble-select and nimble-listbox-option (5454342+brianehenry@users.noreply.github.com)
+
+## 1.0.0-beta.46
+
+Thu, 09 Sep 2021 16:11:55 GMT
+
+### Changes
+
+- add readonly text-field support/styling (22551874+scotia673@users.noreply.github.com)
+
+## 1.0.0-beta.45
+
+Thu, 09 Sep 2021 15:38:58 GMT
+
+### Changes
+
+- use end slot default value for text-field (22551874+scotia673@users.noreply.github.com)
+
+## 1.0.0-beta.44
+
+Thu, 09 Sep 2021 15:24:52 GMT
+
+### Changes
+
+- improve spacing in text-field storybook matrix (22551874+scotia673@users.noreply.github.com)
+
+## 1.0.0-beta.43
+
+Wed, 08 Sep 2021 23:12:35 GMT
+
+### Changes
+
+- Update dependencies (jattasNI@users.noreply.github.com)
+- Bump @ni/nimble-components to v1.0.0-beta.43 (jattasNI@users.noreply.github.com)
+
+## 1.0.0-beta.42
+
+Tue, 07 Sep 2021 20:53:58 GMT
+
+### Changes
+
+- tweak number and text field styles based on UX updates (22551874+scotia673@users.noreply.github.com)
+
+## 1.0.0-beta.41
+
+Tue, 07 Sep 2021 20:01:48 GMT
+
+### Changes
+
+- hide password reveal icon in Edge (22551874+scotia673@users.noreply.github.com)
+
+## 1.0.0-beta.40
+
+Fri, 03 Sep 2021 21:30:56 GMT
+
+### Changes
+
+- fix icon issue in firefox (22551874+scotia673@users.noreply.github.com)
+
+## 1.0.0-beta.39
+
+Fri, 03 Sep 2021 21:06:53 GMT
+
+### Changes
+
+- Component usage docs strategy (1458528+fredvisser@users.noreply.github.com)
+
+## 1.0.0-beta.38
+
+Fri, 03 Sep 2021 16:05:47 GMT
+
+### Changes
+
+- [object Object] (26874831+atmgrifter00@users.noreply.github.com)
 
 ## 1.0.0-beta.37